#41 Created entity snapshot and data access context
