+++ Tigernet.Samples.RestApi/Clevers/Interfaces/IUserEntityManager.cs
@@ -1,0 +1,11 @@
+﻿using Tigernet.Hosting.DataAccess.EntityManager;
+using Tigernet.Samples.RestApi.Models;
+
+namespace Tigernet.Samples.RestApi.Clevers.Interfaces;
+
+public interface IUserEntityManager : IEntityManagerBase<User>
+{
+    User Add(User user);
+
+    User Update(User user);
+}
