+++ Tigernet.Samples.RestApi/Models/User.cs
@@ -1,11 +1,11 @@
 ﻿using Tigernet.Hosting.Attributes.Query;
-using Tigernet.Hosting.Models.Common;
+using Tigernet.Hosting.DataAccess.Entity;
 
 namespace Tigernet.Samples.RestApi.Models
 {
-    public class User : IEntity<int>, IQueryableEntity
+    public class User : IEntity, IQueryableEntity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         [SearchableProperty]
         public string Name { get; set; }
