+++ src/Tigernet.Hosting/DataAccess/DataSourceProvider/DataSourceProvider.cs
@@ -1,0 +1,6 @@
+namespace Tigernet.Hosting.DataAccess.DataSourceProvider;
+
+public class DataSourceProvider
+{
+    
+}
