+++ src/Tigernet.Hosting/DataAccess/Entity/IEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace Tigernet.Hosting.DataAccess.Entity;
+
+public interface IEntity
+{
+    long Id { get; set; }
+}
