+++ src/Tigernet.Hosting/DataAccess/Entity/IQueryableEntity.cs
@@ -1,0 +1,9 @@
+﻿namespace Tigernet.Hosting.DataAccess.Entity
+{
+    /// <summary>
+    /// Defines queryable entity
+    /// </summary>
+    public interface IQueryableEntity
+    {
+    }
+}
