+++ src/Tigernet.Hosting/DataAccess/EntityChangeTracker/EntityState.cs
@@ -1,0 +1,9 @@
+namespace Tigernet.Hosting.DataAccess.EntityChangeTracker;
+
+public enum EntityState
+{
+    Unchanged,
+    Added,
+    Deleted,
+    Modified
+}
