+++ src/Tigernet.Hosting/DataAccess/QueryOptions/FilterOptions.cs
@@ -1,0 +1,19 @@
+﻿namespace Tigernet.Hosting.DataAccess.QueryOptions;
+
+/// <summary>
+/// Represents filtering options
+/// </summary>
+/// <typeparam name="TModel"></typeparam>
+public class FilterOptions<TModel> where TModel : class
+{
+    public FilterOptions() => (Filters) = new List<QueryFilter>();
+
+    public FilterOptions(IEnumerable<QueryFilter> filters) => (Filters) = filters.ToList();
+
+    public List<QueryFilter> Filters { get; set; }
+
+    public QueryFilter this[string key]
+    {
+        get { return Filters.FirstOrDefault(x => x.Key == key); }
+    }
+}
