+++ src/Tigernet.Hosting/DataAccess/QueryOptions/IEntityQueryOptions.cs
@@ -1,0 +1,15 @@
+﻿using Tigernet.Hosting.DataAccess.Entity;
+
+namespace Tigernet.Hosting.DataAccess.QueryOptions;
+
+/// <summary>
+/// Defines properties for queryable entities source query options
+/// </summary>
+/// <typeparam name="TEntity"></typeparam>
+public interface IEntityQueryOptions<TEntity> : IQueryOptions<TEntity> where TEntity : class, IQueryableEntity
+{
+    /// <summary>
+    /// Requested include model options
+    /// </summary>
+    IncludeOptions<TEntity>? IncludeOptions { get; set; }
+}
