--- src/Tigernet.Hosting/Models/Common/IEntity.cs
@@ -1,6 +1,0 @@
-﻿namespace Tigernet.Hosting.Models.Common;
-
-public interface IEntity<TKey>
-{
-    TKey Id { get; set; }
-}
