--- src/Tigernet.Hosting/Models/Common/IQueryableEntity.cs
@@ -1,11 +1,0 @@
-﻿using Tigernet.Hosting.Models.Query;
-
-namespace Tigernet.Hosting.Models.Common
-{
-    /// <summary>
-    /// Defines queryable entity
-    /// </summary>
-    public interface IQueryableEntity
-    {
-    }
-}
