--- src/Tigernet.Hosting/Models/Query/QueryFilter.cs
@@ -1,19 +1,0 @@
-﻿namespace Tigernet.Hosting.Models.Query;
-
-/// <summary>
-/// Represents queryable source query options
-/// </summary>
-public class QueryFilter
-{
-    public QueryFilter(string key, string? value) => (Key, Value) = (key, value);
-
-    /// <summary>
-    /// Field key name
-    /// </summary>
-    public string Key { get; }
-
-    /// <summary>
-    /// Filtering value
-    /// </summary>
-    public string? Value { get; }
-}
