--- src/Tigernet.Hosting/Models/Query/QueryOptions.cs
@@ -1,23 +1,0 @@
-﻿using Tigernet.Hosting.Models.Query;
-
-namespace Tigernet.Hosting.Models.Query;
-
-/// <summary>
-/// Represents queryable source query options
-/// </summary>
-/// <typeparam name="TModel"></typeparam>
-public class QueryOptions<TModel> : IQueryOptions<TModel> where TModel : class
-{
-    public QueryOptions()
-    {
-        PaginationOptions = new PaginationOptions(10, 1);
-    }
-
-    public SearchOptions<TModel>? SearchOptions { get; set; }
-
-    public FilterOptions<TModel>? FilterOptions { get; set; }
-
-    public SortOptions<TModel>? SortOptions { get; set; }
-
-    public PaginationOptions PaginationOptions { get; set; }
-}
